+++ test/test.search.js
@@ -108,6 +108,12 @@ describe('Search', function() {
       var result = list.search('jonny', [ 'pet' ]);
       expect(result.length).to.equal(0);
     });
+    it('should remove columnm option', function() {
+      var result = list.search('jonny', [ 'born' ]);
+      expect(result.length).to.equal(0);
+      var result = list.search('jonny');
+      expect(result.length).to.equal(1);
+    });
   });
 
   describe('Custom search function', function() {